requirements.txt
langchain_ollama
langchain
argparse
moviepy
openai-whisper